Add support for Nimble packages
This gives the playground support to have Nimble packages installed. The
list of packages installed will be the same for all running instances.
If you want to add a package to the currently very short list, make a PR
that adds it as a dependency in the packages.nimble file.