Add ability to choose version
This adds the ability to choose the version used to compile the code.
The versions are grabbed from the tags of the Nim repository, with some
releases ignored as they still used "nimrod" as the name of the
executable. The Docker image now supports building all the various
versions, and the updateimages.sh script will create and tag a container
for every version not found by docker. Only the latest version will be
built with nimble packages, for now that means only v1.0.0, but will
include every image from now unless they are manually deleted. The
back-end now also has a new route /versions that returns every supported
version. And the JSON object given to /compile now has a new field
"version" that takes an entry from this list, or the literal "latest"
and uses that version.